backend/pyproject.toml
[project]
name = "ai-cmo"
version = "0.1.0"
description = "AiCMO backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "celery>=5.5.1",
    "fastapi>=0.115.12",
    "google-cloud-aiplatform>=1.86.0",
    "google-genai>=1.20.0",
    "litellm<=1.76.0",
    "psycopg[binary]>=3.2.6",
    "pydantic-extra-types>=2.10.3",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
    "valkey[libvalkey]>=6.1.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
improved-auth = {path = "../shared/improved_auth", editable = true}


[tool.basedpyright]
include=["."]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.5",
]
